[REFACTOR] GenrePreference 리포지토리 패키지 변경
domain -> repository
